requirements.txt
pyserial
av
opencv-python
PySide6
numpy
Pillow